container/requirements.txt
flask
flask-socketio
flask-login
simple-websocket
python-pam
